refactor(artist): move album card into widgets
shared with discography page, fills widht it is give so it works in both
rail and grid
